more stuff

+++ NewGame/scripts/scenes/inventory.lua
@@ -86,9 +86,9 @@ function move_inv()
         if (not inv_ob("up")) inv_pos[2]-=1
     elseif btnp(⬇️) then
         if (not inv_ob("down")) inv_pos[2]+=1
-    elseif btnp(❎) then
+    elseif btnp(❎) and inventory[pos] != nil then
         inv_prompt=true
-    elseif btnp(🅾️) then
+    elseif btnp(🅾️) and inventory[pos] != nil then
         delete_item(pos)
     end
 end
